refactoring migrated Endpoint to protocol package
